+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Suggested subheadings: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Se
 
 If there are not many items, just list them sequentially. 
 
+# Unreleased
+
 * Fix: UI Clean Up: "Add Tags" should not be tappable (#610)