+++ CHANGELOG.md
@@ -13,6 +13,7 @@ If there are not many items, just list them sequentially.
 # Unreleased
 
 * Fix: UI Clean Up: "Add Tags" should not be tappable (#610)
+* UI Clean Up: Remove pressed state for Cover Image within Bot Profile (#615)
 
 
 # 1.28.1 - 2017 April 28