+++ CHANGELOG.md
@@ -12,10 +12,10 @@ If there are not many items, just list them sequentially.
 
 # Unreleased
 
-* Fix: UI Clean Up: "Add Tags" should not be tappable (#610)
+* UI Clean Up: "Add Tags" should not be tappable (#610)
 * UI Clean Up: Remove pressed state for Cover Image within Bot Profile (#615)
 * UI Clean Up: Implement an "x" during Share Flow when user types (#614)
-* Fix: UI Clean Up: Remove "Follow Me" CTA (#612)
+* UI Clean Up: Remove "Follow Me" CTA (#612)
 * UI Clean Up: Remove "Add Cover Image" Icon from part of the Bot Create flow (#611)
 * More descriptive error messages for image upload (to help with #558)